+++ src/ethereum/info/amoyInfo.ts
@@ -41,6 +41,7 @@ const networkFees: EthereumFees = {
 }
 
 const networkInfo: EthereumNetworkInfo = {
+  addressQueryLookbackBlocks: 4, // 2 minutes
   networkAdapterConfigs: [
     {
       type: 'rpc',