+++ src/ethereum/info/ethDevInfo.ts
@@ -101,6 +101,7 @@ const networkFees: EthereumFees = {
 }
 
 const networkInfo: EthereumNetworkInfo = {
+  addressQueryLookbackBlocks: 8, // 2 minutes
   networkAdapterConfigs: [
     {
       type: 'rpc',