+++ src/ethereum/info/ethereumpowInfo.ts
@@ -55,6 +55,7 @@ const networkFees: EthereumFees = {
 }
 
 const networkInfo: EthereumNetworkInfo = {
+  addressQueryLookbackBlocks: 8, // 2 minutes
   networkAdapterConfigs: [
     {
       type: 'rpc',