+++ src/ethereum/info/fantomInfo.ts
@@ -283,6 +283,7 @@ const networkFees: EthereumFees = {
 
 // Exported for fee provider test
 const networkInfo: EthereumNetworkInfo = {
+  addressQueryLookbackBlocks: 60, // 2 minutes
   networkAdapterConfigs: [
     {
       type: 'rpc',