+++ src/ethereum/info/sepoliaInfo.ts
@@ -46,6 +46,7 @@ const networkFees: EthereumFees = {
 }
 
 const networkInfo: EthereumNetworkInfo = {
+  addressQueryLookbackBlocks: 8, // 2 minutes
   networkAdapterConfigs: [
     {
       type: 'rpc',